requirements.txt
asgiref==3.4.1
Django==3.2.7
pkg_resources==0.0.0
psycopg2==2.9.1
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.2
djangorestframework-simplejwt==4.8.0
